requirements.txt
FrEIA==0.2
librosa==0.10.1
matplotlib==3.8.0
numpy==1.26.4
PyYAML==6.0.1
resampy==0.4.3
scipy==1.12.0
soundfile==0.12.1
speechbrain==1.0.0
torch==1.13.1
torchaudio==0.13.1
